Changement duree

+++ include/GPCalendar/Model/rendezvous.h
@@ -20,7 +20,7 @@ typedef struct{
     char * motif;
 }RendezVous;
 
-RendezVous * CreerRendezVous(int an, int mois, int jour, double heure_debut, int duree, char * lieu, Patient * patient, Medecin * medecin, char * motif);
+RendezVous * CreerRendezVous(int an, int mois, int jour, double heure_debut, double duree, char * lieu, Patient * patient, Medecin * medecin, char * motif);
 void FreeRendezVous(RendezVous * rdv);
 int AnnulerRendezVous(RendezVous * rdv);
 RendezVous * DeplacerRendezVous(RendezVous * rdv, int n_an, int n_mois, int n_jour, double n_heure_debut, int n_duree);
